DDA-259: Refactor mock connectors
