+++ src/react/client/mockResponses/GET_DIORY_FOLDER_LOCATION.json
@@ -1,3 +1,3 @@
 {
-  "dioryFolderLocation": "public/diory-demo-content"
+  "dioryFolderLocation": "public/diory"
 }
