+++ src/react/connectors/mock/GET_DIOSPHERE.json
