+++ src/react/connectors/mock/mockConnector.js
@@ -1,4 +1,4 @@
-import GET_DIOGRAPH from './mockResponses/GET_DIOGRAPH.json'
+import GET_DIOGRAPH from './GET_DIOGRAPH.json'
 
 export function getDiograph(params) {
   const mockResponse = GET_DIOGRAPH
@@ -10,6 +10,13 @@ export function getDiograph(params) {
 export function saveDiograph(params) {
   const mockResponse = true
   console.log('mock: saveDiograph', params, mockResponse)
+
+  return mockResponse
+}
+
+export function getDiosphere(params) {
+  const mockResponse = GET_DIOGRAPH
+  console.log('mock: getDiosphere', params, mockResponse)
 
   return mockResponse
 }
